Fix use-after-frees/double-frees of temporaries
Every instance of the expression .to_cstring().unwrap().as_ptr() creates
an unnamed (temporary) CString, then grabs a pointer to it. Before this
pointer is used, the allocation that the CString owns is freed. So using
this pointer at all is always a use after free. Some of these pointers
appear to never be read from, in those cases this is a double free.

This can be discovered by compiling and running this crate with
RUSTFLAGS=-Zsanitizer=address. I did not debug each instance of this
individually, I noticed the pattern (which is depressingly common) and
did a find-and-replace to change the as_ptr() in that case to
into_raw(), which leaks the allocation (or returns an owning raw pointer,
depending on how you look at it). With this change, AddressSanitizer no
longer reports any problems.